BUG: fftpack: use single-precision FFT implementation only for "easy" sizes
For large "difficult" sizes, the current single precision implementation
apparently falls back to a non-optimal algorithm, and can produce large
rounding errors. (bug #1212)

This commit provides a work-around, by enabling the single-precision
algorithms only for sizes that are composite numbers of 2, 3, and 5, for
which FFTPACK has explicit support. This is probably a rather
conservative approach.

git-svn-id: http://svn.scipy.org/svn/scipy/branches/0.8.x@6570 d6536bca-fef9-0310-8506-e4c0a848fbcf